% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_and}
\alias{add_and}
\title{Add UND or AND to Listing}
\usage{
add_and(words, sep = ", ", and = "und")
}
\description{
Add UND or AND before the last element of a character vector.
}
\examples{
\dontrun{add_and(LETTERS[1:3])}
}
