% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cal}
\alias{cal}
\title{{Print calendar to console}}
\source{
{\url{https://www.garrickadenbuie.com/blog/r-console-calendar/}}
}
\usage{
cal(
  start_date = lubridate::today(),
  end_date = start_date + 28,
  week_start = 1
)
}
\description{
{Prints a calendar to the console / RMarkdown documents}
}
\author{
{Garrick Aden-Buie}
}
