% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{exp_labs}
\alias{exp_labs}
\title{{Scale labels after exponential transformation}}
\usage{
exp_labs(x, digits = 2)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{Number of digits}
}
\description{
{Scale labels for ggplot2 graphics after 
exponential transformation}
}
