% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_num}
\alias{format_num}
\title{Format numbers.}
\usage{
format_num(x, digits = 1)
}
\arguments{
\item{x}{number}

\item{digits}{number of digits}
}
\description{
Format numbers (in R Markdown documents)
}
\examples{
\dontrun{
format_num(10.2589, digits = 2)
}
}
