% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{formatNum}
\alias{formatNum}
\title{Format numbers.}
\usage{
formatNum(x, digits = 1)
}
\arguments{
\item{x}{number}

\item{digits}{number of digits}
}
\description{
Format numbers (in R Markdown documents)
}
\examples{
\dontrun{
formatNum(10.2589, digits = 2)
}
}
