% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{ggnice_hist}
\alias{ggnice_hist}
\title{{Plot histogram with some additional statistics using ggplot2}}
\usage{
ggnice_hist(
  data,
  x,
  x_lab = NULL,
  title = NULL,
  vline = NULL,
  digits = 1,
  scale = scales::identity_trans(),
  theme = mdthemes::md_theme_bw()
)
}
\arguments{
\item{data}{data.frame}

\item{x}{numeric variable (part of data)}

\item{x_lab}{Label for x-axis (character)}

\item{title}{Plot title (character)}

\item{vline}{x-intercept for vertical line (numeric)}

\item{digits}{Number of digits (numeric)}

\item{scale}{Scale transformation (character); defaults to "identity" (no transformation)}

\item{theme}{ggplot-theme (function)}
}
\description{
{Plot histogram with some additional statistics using ggplot2.} 
Ported from the OurTools package written by Dirk Hasenclever.
}
\examples{
\dontrun{
ggnice_hist(data = mtcars, 
             x = hp, 
             x_lab = 'Gross horsepower',
             title = 'mtcars: Gross horsepower',
             scale = 'identity',
             theme = mdthemes::md_theme_bw())
}
}
