% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{ggnorm_dist}
\alias{ggnorm_dist}
\title{{Plot normal distribution using ggplot2}}
\usage{
ggnorm_dist(
  line_color = "white",
  fill_color = "steelblue",
  arrow_color = "black"
)
}
\arguments{
\item{line_color}{Colour of the vertical lines}

\item{fill_color}{Colour of the area under the curve}

\item{arrow_color}{Colour of the arrows}
}
\description{
{Plots normal distribution along with standard deviations 
and confidence intervals using ggplot2.}
}
