% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{globenv_2_list}
\alias{globenv_2_list}
\title{{Put objects of global environment into a list}}
\usage{
globenv_2_list(pattern = "^df")
}
\arguments{
\item{pattern}{Pattern to search for; may be a 
regular expression}
}
\description{
{Put objects of global environment 
matching a pre-defined pattern into a list}
}
