% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ht}
\alias{ht}
\title{{Display the top and bottom of a data frame}}
\usage{
ht(data, n = 5)
}
\arguments{
\item{data}{data frame}

\item{n}{number of rows to display from head and tail
\dontrun{
ht(mtcars)
}}
}
\description{
{Quickly display the top and bottom of a data frame}
}
