% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{increase_decrease}
\alias{increase_decrease}
\title{Return 'increased' vs. 'decreased' if value is positive vs. negative.}
\usage{
increase_decrease(x)
}
\arguments{
\item{x}{number}
}
\description{
Return 'increased' vs. 'decreased' if value is positive vs. negative (in R Markdown documents)
}
\examples{
\dontrun{
increase_decrease(-0.2)
}
}
