% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{model_2_txt}
\alias{model_2_txt}
\title{{Print estimate, 95\% CI, and p-value of regression modles}}
\usage{
model_2_txt(
  data,
  beta = estimate,
  unit = "",
  lower = conf.low,
  upper = conf.high,
  p = p.value,
  term,
  filter_var,
  filter_value,
  digits = 2
)
}
\arguments{
\item{data}{data frame as derived from broom::tidy(fit)}

\item{beta}{estimate}

\item{unit}{unit of the estimate (needs to be quoted)}

\item{lower}{Lower bound of 95\% CI}

\item{upper}{Upper bound of 95\% CI}

\item{p}{p-value}

\item{term}{Name of independent variable (needs to be quoted, partial match is sufficient)}

\item{filter_var}{Name of variable for which a filter should be applied}

\item{filter_value}{Term to be filtered (needs to be quoted)}

\item{digits}{Number of digits for estimate and CI (default: 2)

\dontrun{
model_2_txt(df)
}}
}
\description{
{Print estimate, 95\% CI, and p-value as derived from broom::tidy(fit)}
}
