% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{modulus_trans}
\alias{modulus_trans}
\title{{Modulus scale transformation}}
\source{
{\url{http://freerangestats.info/blog/2015/09/05/creating-a-scale-transformation}}
}
\usage{
modulus_trans(lambda)
}
\arguments{
\item{lambda}{Tuning parameter}
}
\description{
{Modulus scale transformation function for ggplot2}
}
