% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{more_less}
\alias{more_less}
\title{{Return 'more' vs. 'less' if value is positive vs. negative}}
\usage{
more_less(x, rev = FALSE)
}
\arguments{
\item{x}{number}

\item{rev}{FALSE (default) for increasing trend, TRUE for decreasing trend (reduction).}
}
\description{
{Return 'more' vs. 'less' if value is positive vs. negative 
(in R Markdown documents)}
}
