% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{n2w_de}
\alias{n2w_de}
\title{{Convert numbers to German words}}
\usage{
n2w_de(x, cap = FALSE, female = FALSE)
}
\arguments{
\item{x}{A numeric vector. Values should be integers. The absolute values should be between 0 and 10.}

\item{cap}{Whether to capitalize the first letter of the word. This can be useful when the word is at the 
beginning of a sentence. Default is FALSE.}

\item{female}{Logical. Make gender of "ein" female ("eine").}
}
\description{
{Convert numbers from 0 to 10 to German words.}
}
