% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{named_group_split}
\alias{named_group_split}
\title{{Named {group_split()} function}}
\usage{
named_group_split(.tbl, ...)
}
\arguments{
\item{.tbl}{Grouping variable}
}
\description{
{Modified version of {dplyr::group_split()}. 
Unlike the original function, it returns named lists.}
}
