% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{paste_lipsum}
\alias{paste_lipsum}
\title{{Print 'Lorem Ipsum'}}
\source{
{\url{https://github.com/aakosm/lipsum}}
}
\usage{
paste_lipsum(n = 1)
}
\arguments{
\item{n}{The number of lorem ipsum paragraphs to be printed}
}
\description{
{Print 'Lorem Ipsum' to the console or
in R Markdown documents}
}
