% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_by_2sd}
\alias{scale_by_2sd}
\title{{Scaling by two standard deviations}}
\usage{
scale_by_2sd(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
{Scale a numeric vector by by two standard deviations
(rather than by one)}
}
