% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{scale_x_ln2}
\alias{scale_x_ln2}
\title{{Logarithmic transformation of x-scale}}
\usage{
scale_x_ln2(...)
}
\description{
{Logarithmic transformation of x-scale in ggplot graphics.}
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(x = mpg)) + geom_density() + scale_x_ln2()
}
}
