% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{scale_y_ln2}
\alias{scale_y_ln2}
\title{Logarithmic transformation of y-scale}
\usage{
scale_y_ln2(...)
}
\description{
Logarithmic transformation of y-scale in ggplot graphics.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(mpg)) + geom_density() + scale_y_ln2()
}
}
