% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctcl.R
\name{sect_map_varlabs}
\alias{sect_map_varlabs}
\title{{Secutrial: Mapping variable names to variable labels}}
\usage{
sect_map_varlabs(modul)
}
\arguments{
\item{modul}{Name of table from SecuTrial database}
}
\value{
The function returns a tibble with 2 variables:
 \itemize{
 \item \code{ITEM_COLUMN:} {Variable name}
 \item \code{ITEM_LABEL:} {Variable label}
}
}
\description{
{Mapping the variable names of a table from SecuTrial to variable labels}
}
\examples{
\dontrun{
df <- map_varlabs(modul = 'mnppal_heart')
}
}
