% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_in_excel}
\alias{show_in_excel}
\title{{Show data.frame in MS Excel}}
\source{
{\url{https://twitter.com/CorradoLanera/status/1447478650488737792}}
}
\usage{
show_in_excel(.data)
}
\arguments{
\item{.data}{data.frame}
}
\description{
{Show data.frame in MS Excel and continue piping}
}
