% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{sliding_incidence_rate}
\alias{sliding_incidence_rate}
\title{{Sliding Incidence Rate}}
\usage{
sliding_incidence_rate(
  data,
  var_start_date = ICU_START,
  var_end_date = ICU_END,
  var_event_date = MRE_DATE,
  var_grouping = PERIOD,
  offset = list(start = 3, end = 2),
  time_span = 365,
  scale_fct = 1000,
  window = 30,
  step = 1
)
}
\arguments{
\item{data}{{A data.frame }}

\item{var_start_date}{{Variable with start dates (start of episode)}}

\item{var_end_date}{{Variable with end dates (end of episode)}}

\item{var_event_date}{{Variable with date of event}}

\item{var_grouping}{{Grouping variable (default: Intervention period)}}

\item{offset}{{
\itemize{
 \item \code{start} {Days between start of period and beginnung of time under risk}
 \item \code{end} {Days between end of period and end of time under risk}
}
}}

\item{time_span}{{Length of intervention period (in days)}}

\item{scale_fct}{{Scaling factor for incidence rate}}

\item{window}{{Size of sliding window (in days)}}

\item{step}{{Step between sliding windows (in days)}}
}
\value{
{A data.frame with sliding incidence rates}
}
\description{
{Calculate incidence rate with sliding windows}
}
\author{
{Dirk Hasenclever, Norbert Koehler}
}
