% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{sliding_IR}
\alias{sliding_IR}
\title{{Sliding Incidence Rate}}
\usage{
sliding_IR(
  data,
  var_start_date = ICU_START,
  var_end_date = ICU_END,
  var_event_date = NULL,
  var_grouping = PERIOD,
  offset = list(start = 3, end = 2),
  time_span = 365,
  scale_fct = 1000,
  window = 20,
  step = 3
)
}
\arguments{
\item{data}{{A data.frame }}

\item{var_start_date}{{Variable with first day of episodes (default: ICU_START)}}

\item{var_end_date}{{Variable with last day of episodes (default: ICU_END))}}

\item{var_event_date}{{Variable with date of event (NULL)}}

\item{var_grouping}{{Grouping variable (default: PERIOD)}}

\item{offset}{{
\itemize{
 \item \code{start} {Number of days between start of period 
 and beginning of time under risk (default: 3)}
 \item \code{end} {Number of days between end of period 
 and end of time under risk (default: 2)}
}
}}

\item{time_span}{{Length of intervention period in days (default: 365)}}

\item{scale_fct}{{Scaling factor for incidence rate (default: 1.000)}}

\item{window}{{Size of sliding window in days (default: 20)}}

\item{step}{{Step between sliding windows in days (default: 3)}}
}
\value{
{A data.frame with sliding incidence rates}
}
\description{
{Calculate incidence rate with sliding windows}
}
\author{
{Dirk Hasenclever, Norbert Koehler}
}
