% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{str_2_df}
\alias{str_2_df}
\title{{Tranform a character vector into a data frame}}
\usage{
str_2_df(x, n_col = 6, order = "horizontal")
}
\arguments{
\item{x}{Character vector}

\item{n_col}{Desired number of columns for the data frame}

\item{order}{Shall the character vector be ordered horizontally 
('horizontal', default) or vertically ('vertical')?}
}
\description{
{A character vector can be printed into a bulleted list.
However, a long list may be rather space consuming. This function transforms
a character vector into a data frame, which may be printed as a table.}
}
