% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{str_abbrev}
\alias{str_abbrev}
\title{{Shorten character string}}
\usage{
str_abbrev(x, l_max = 10)
}
\arguments{
\item{x}{character string to be shortened}

\item{l_max}{maximal length of character string (numeric), default = 10}
}
\description{
{Shortens a character string to a specified length}
}
\examples{
\dontrun{str_abbrev('Hypertension', l_max = 6)}
}
