% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{ggnorm_dist}
\alias{ggnorm_dist}
\title{{Plot normal distribution using ggplot2}}
\source{
{\url{https://github.com/svmiller/stevemisc}}
}
\usage{
ggnorm_dist(
  curvecolor = "grey10",
  fillcolor = "#C7271C",
  arrowcolor = "yellow"
)
}
\arguments{
\item{curvecolor}{Colour of the curve}

\item{fillcolor}{Colour of the area under the curve}

\item{arrowcolor}{Colour of the arrows}
}
\description{
{Plot normal distribution along with standard deviations 
and confidence intervals using ggplot2.} 
Modified version of stevemisc::normal_distribution()
}
\examples{
\dontrun{
ggnorm_dist()
}
}
